Массив с данными вынесен в отдельный файл
