+++ src/components/main/centerblock/contentblock/trackblock/trackblock.jsx
@@ -1,28 +1,15 @@
 import React from 'react';
 import "./trackblock.css";
 import Skeleton from '../../../../skeleton/skeleton';
+import { workArray } from '../../../../../workarray';
+
+
+const trackInfo=workArray();
+
 
 function Trackblock({loading}){
     
-    const trackInfo =[
-      { id:1, 'track title': 'Guild','track comment':'','track author': 'Nero','track album': 'Welcome Reality', 'track time': '4:44'},
-      { id:2, 'track title': 'Elektro','track comment':'','track author': 'Dynoro, Outwork, Mr. Gee','track album': 'Elektro', 'track time': '2:22'},
-      { id:3, 'track title': 'I’m Fire','track comment':'','track author': 'Ali Bakgor','track album': 'I’m Fire', 'track time': '2:22'},
-      { id:4, 'track title': 'Non Stop','track comment':'(Remix)','track author': 'Стоункат, Psychopath','track album': 'Non Stop', 'track time': '4:12'},
-      { id:5, 'track title': 'Run Run','track comment':'(feat. AR/CO)','track author': 'Jaded, Will Clarke, AR/CO','track album': 'Run Run', 'track time': '2:54'},
-      { id:6, 'track title': 'Eyes on Fire','track comment':'(Zeds Dead Remix)','track author': 'Blue Foundation, Zeds Dead','track album': 'Eyes on Fire', 'track time': '5:20'},
-      { id:7, 'track title': 'Mucho Bien','track comment':'(Hi Profile Remix)','track author': 'HYBIT, Mr. Black, Offer Nissim, Hi Profile','track album': 'Mucho Bien', 'track time': '3:41'},
-      { id:8, 'track title': 'Knives n Cherries','track comment':'','track author': 'minthaze','track album': 'Captivating', 'track time': '1:48'},
-      { id:9, 'track title': 'Knives n Cherries','track comment':'','track author': 'minthaze','track album': 'Captivating', 'track time': '1:48'},
-      { id:10, 'track title': 'Knives n Cherries','track comment':'','track author': 'minthaze','track album': 'Captivating', 'track time': '1:48'},
-      { id:11, 'track title': 'Knives n Cherries','track comment':'','track author': 'minthaze','track album': 'Captivating', 'track time': '1:48'},
-      { id:12, 'track title': 'Knives n Cherries','track comment':'','track author': 'minthaze','track album': 'Captivating', 'track time': '1:48'},
-      { id:13, 'track title': 'Knives n Cherries','track comment':'','track author': 'minthaze','track album': 'Captivating', 'track time': '1:48'},
-      { id:14, 'track title': 'Knives n Cherries','track comment':'','track author': 'minthaze','track album': 'Captivating', 'track time': '1:48'},
-      { id:15, 'track title': 'How Deep Is Your Love','track comment':'','track author': 'Calvin Harris, Disciples','track album': 'How Deep Is Your Love', 'track time': '3:32'},
-      { id:16, 'track title': 'Morena','track comment':'','track author': 'Tom Boxer','track album': 'Soundz Made in Romania', 'track time': '3:36'},
-      { id:17, 'track title': 'kkk','track comment':'kkk','track author': 'kkk','track album': 'kkk', 'track time': 'kkk'},
-    ];
+    
 
     const listItems = trackInfo.map((elem) =>
   <div className="playlist__item" key={elem.id}>
